Modify solo home runs and same-base thrown out calls
For solo home runs, the games were scored explicitly listing 1 RBI,
but the hit function automatically accounts for solo homers. All solo
home run calls have been modified to not list the RBI count.

For thrown_out calls, in cases like pickoffs and double plays where
a runner was out of base, technically the out occurs at the base where
the player stats. This was not supported by the thrown_out method, so it
was been modified.